Python package updates for command language

+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/collision_box_box_unit.hpp
@@ -167,8 +167,8 @@ inline void runTestTyped(DiscreteContactManager& checker, ContactTestType test_t
   }
   else
   {
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], -0.3, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], 1.0, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], -0.3, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], 1.0, 0.001);
   }
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * -1.0, 0.001);
@@ -251,8 +251,8 @@ inline void runTestTyped(DiscreteContactManager& checker, ContactTestType test_t
     }
     else
     {
-      EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 1.1, 0.001);
-      EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], 1.0, 0.001);
+      EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 1.1, 0.001);
+      EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], 1.0, 0.001);
     }
 
     EXPECT_NEAR(result_vector[0].normal[0], idx[2] * -1.0, 0.001);
@@ -289,8 +289,8 @@ inline void runTestTyped(DiscreteContactManager& checker, ContactTestType test_t
     }
     else
     {
-      EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 1.1, 0.001);
-      EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], 1.0, 0.001);
+      EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 1.1, 0.001);
+      EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], 1.0, 0.001);
     }
 
     EXPECT_NEAR(result_vector[0].normal[0], idx[2] * -1.0, 0.001);
