+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/collision_box_capsule_unit.hpp
@@ -144,8 +144,8 @@ inline void runTest(DiscreteContactManager& checker)
   }
   else
   {
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 0.5, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], -0.05, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 0.5, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], -0.05, 0.001);
   }
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 1.0, 0.001);
@@ -195,8 +195,8 @@ inline void runTest(DiscreteContactManager& checker)
   }
   else
   {
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][2], 0.5, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][2], 0.625, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][2], 0.5, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][2], 0.625, 0.001);
   }
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 0.0, 0.0011);  // FCL Required the bump in tolerance
