+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/collision_octomap_mesh_unit.hpp
@@ -105,11 +105,14 @@ inline void runTest(DiscreteContactManager& checker, const std::string& file_pat
     if (r.link_names[0] != "plane_link")
       idx = 1;
 
-    mesh_vertices_color[static_cast<std::size_t>((*mesh_triangles)[4 * r.subshape_id[idx] + 1])] =
+    mesh_vertices_color[static_cast<std::size_t>(
+        (*mesh_triangles)[static_cast<std::size_t>(4 * r.subshape_id[static_cast<std::size_t>(idx)] + 1)])] =
         Eigen::Vector3i(255, 0, 0);
-    mesh_vertices_color[static_cast<std::size_t>((*mesh_triangles)[4 * r.subshape_id[idx] + 2])] =
+    mesh_vertices_color[static_cast<std::size_t>(
+        (*mesh_triangles)[static_cast<std::size_t>(4 * r.subshape_id[static_cast<std::size_t>(idx)] + 2)])] =
         Eigen::Vector3i(255, 0, 0);
-    mesh_vertices_color[static_cast<std::size_t>((*mesh_triangles)[4 * r.subshape_id[idx] + 3])] =
+    mesh_vertices_color[static_cast<std::size_t>(
+        (*mesh_triangles)[static_cast<std::size_t>(4 * r.subshape_id[static_cast<std::size_t>(idx)] + 3)])] =
         Eigen::Vector3i(255, 0, 0);
   }
